IDAM-256 Refactor login and pw reset containers
