+++ custom-pages/universal-login/package.json
@@ -22,6 +22,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "test:teamcity": "react-scripts test --testResultsProcessor ./node_modules/jest-teamcity-reporter",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -42,6 +43,7 @@
   "devDependencies": {
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
+    "jest-teamcity-reporter": "^0.9.0",
     "node-sass": "^4.12.0"
   }
 }
