+++ custom-pages/universal-login/src/components/App/App.scss
@@ -1,4 +1,4 @@
-@import "../../components/variables";
+@import "../variables";
 
 body {
   font-family: var(--font-base);
