--- custom-pages/universal-login/src/containers/PasswordResetApp/PasswordResetApp.scss
@@ -1,45 +1,0 @@
-@import "../../components/variables";
-
-body {
-  font-family: var(--font-base);
-}
-
-.wrapper {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
-
-.col {
-  height: 100%;
-  padding: 3rem;
-  width: var(--col-width-desktop);
-}
-
-.rightCol {
-  padding-top: 1.4rem !important;
-}
-
-.mainContainer {
-  padding-top: var(--top);
-  min-width: var(--col-width-desktop);
-}
-
-@media (max-width: 700px) {
-  .wrapper {
-    flex-direction: column;
-  }
-  .rightCol,
-  .leftCol {
-    width: 100%;
-    min-width: var(--col-width-mobile);
-    padding-bottom: 0;
-  }
-  .mainContainer {
-    min-width: var(--col-width-mobile);
-  }
-
-  .navigation {
-    width: calc(var(--col-width-mobile) - 5rem);
-  }
-}
