+++ custom-pages/universal-login/src/helpers/__tests__/toggleNavigation.test.js
@@ -1,4 +1,4 @@
-import { showNav, hideNav } from './'
+import { showNav, hideNav } from '../index'
 
 global.document.querySelector = jest.fn(() => ({
   setAttribute: jest.fn(),
