+++ custom-pages/universal-login/src/index.js
@@ -4,8 +4,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter as Router} from "react-router-dom";
 import { Promise } from 'es6-promise'
-import LoginApp from './containers/LoginApp/LoginApp'
-import PasswordResetApp from "./containers/PasswordResetApp/PasswordResetApp";
+import App from './components/App/App'
 
 // if IE8
 global.Promise = global.Promise || Promise;
@@ -14,11 +13,6 @@ global.__DEV__ = global.__DEV__ || document.location.host.indexOf('localhost') >
 const rootElement = document.getElementById("root");
 
 ReactDOM.render(
-  <Router>
-    {process.env.REACT_APP_RENDER === 'login' ?
-      <LoginApp /> :
-      <PasswordResetApp />
-    }
-  </Router>,
+  <App />,
   rootElement
 );
