+++ custom-pages/universal-login/src/services/__tests__/constants.test.js
@@ -1,7 +1,7 @@
 import Enzyme from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
 
-import authOpts from './constants'
+import authOpts from '../constants'
 
 Enzyme.configure({ adapter: new Adapter() })
